Move all native methods into separate library
