+++ NativeLib/Linux/Input/Button.cs
@@ -1,0 +1,11 @@
+﻿namespace NativeLib.Linux
+{
+    public enum Button : uint
+    {
+        LEFT = 1,
+        MIDDLE = 2,
+        RIGHT = 3,
+        BACKWARD = 8,
+        FORWARD = 9
+    }
+}
