+++ NativeLib/OSX/Input/CGMouseButton.cs
@@ -1,0 +1,11 @@
+﻿namespace NativeLib.OSX
+{
+    public enum CGMouseButton : uint
+    {
+        kCGMouseButtonLeft = 0,
+        kCGMouseButtonRight = 1,
+        kCGMouseButtonCenter = 2,
+        kCGMouseButtonBackward = 3,
+        kCGMouseButtonForward = 4,
+    }
+}
