+++ NativeLib/Windows/Generic/POINT.cs
@@ -1,0 +1,17 @@
+using System.Runtime.InteropServices;
+
+namespace NativeLib.Windows
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct POINT
+    {
+        public int X;
+        public int Y;
+
+        public POINT(int x, int y)
+        {
+            this.X = x;
+            this.Y = y;
+        }
+    }
+}
