+++ NativeLib/Windows/Generic/Rect.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.InteropServices;
+
+namespace NativeLib.Windows
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct Rect
+    {
+        public int left;
+        public int top;
+        public int right;
+        public int bottom;
+    }
+}
