+++ NativeLib/Windows/Input/KEYSTATE.cs
@@ -1,0 +1,8 @@
+namespace NativeLib.Windows.Input
+{
+    public enum KEYSTATE : int
+    {
+        KEY_TOGGLED = 0x1,
+        KEY_PRESSED = 0x8000
+    }
+}
