+++ NativeLib/Windows/Input/VirtualKeyStates.cs
@@ -1,0 +1,11 @@
+﻿namespace NativeLib.Windows.Input
+{
+    public enum VirtualKeyStates : int
+    {
+        VK_LBUTTON = 0x01,
+        VK_RBUTTON = 0x02,
+        VK_MBUTTON = 0x04,
+        VK_XBUTTON1 = 0x05,
+        VK_XBUTTON2 = 0x06,
+    }
+}
