+++ NativeLib/Windows/Input/XBUTTON.cs
@@ -1,0 +1,9 @@
+﻿namespace NativeLib.Windows
+{
+    public enum XBUTTON : uint
+    {
+        NONE = 0x0000,
+        XBUTTON1 = 0x0001,
+        XBUTTON2 = 0x0002
+    }
+}
