+++ OpenTabletDriver.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TabletDriverLib", "TabletDr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTabletDriverGUI", "OpenTabletDriverGUI\OpenTabletDriverGUI.csproj", "{5720DD79-C5B1-410A-B059-AC385F702DA5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NativeLib", "NativeLib\NativeLib.csproj", "{57A5101A-20F4-41AA-8649-2C3785DD5BA2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,5 +46,17 @@ Global
 		{5720DD79-C5B1-410A-B059-AC385F702DA5}.Release|x64.Build.0 = Release|Any CPU
 		{5720DD79-C5B1-410A-B059-AC385F702DA5}.Release|x86.ActiveCfg = Release|Any CPU
 		{5720DD79-C5B1-410A-B059-AC385F702DA5}.Release|x86.Build.0 = Release|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Debug|x64.Build.0 = Debug|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Debug|x86.Build.0 = Debug|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Release|x64.ActiveCfg = Release|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Release|x64.Build.0 = Release|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Release|x86.ActiveCfg = Release|Any CPU
+		{57A5101A-20F4-41AA-8649-2C3785DD5BA2}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
