+++ TabletDriverLib/Interop/Display/MacOSDisplay.cs
@@ -1,17 +1,21 @@
+using NativeLib.OSX;
+
 namespace TabletDriverLib.Interop.Display
 {
+    using static OSX;
+
     public class MacOSDisplay : IDisplay
     {
-        private uint MainDisplay => Native.MacOSX.CGMainDisplayID();
+        private uint MainDisplay => CGMainDisplayID();
 
         public float Width 
         {
-            get => Native.MacOSX.CGDisplayPixelsWide(MainDisplay);
+            get => CGDisplayPixelsWide(MainDisplay);
         }
 
         public float Height
         {
-            get => Native.MacOSX.CGDisplayPixelsHigh(MainDisplay);
+            get => CGDisplayPixelsHigh(MainDisplay);
         }
     }
 }
