+++ TabletDriverLib/TabletDriverLib.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="System.Numerics.Vectors" Version="4.5.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\NativeLib\NativeLib.csproj" />
+  </ItemGroup>
+
 </Project>
